requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
requests>=2.31.0

# AI / LLM
langchain>=0.1.0,<0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
anthropic>=0.15.0

# Web Scraping & Search
beautifulsoup4>=4.12.0
playwright>=1.41.0
html2text>=2024.2.26
duckduckgo-search>=5.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Scheduling
apscheduler>=3.10.0

# Document Generation
jinja2>=3.1.0
markdown>=3.5.0
weasyprint>=60.0

# API / Web Server (for Gatekeeper UI/API)
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
